requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0
tweepy>=4.14.0
pyyaml>=6.0.2
python-dateutil>=2.9.0
